更新droplet-lib
**复现步骤**

**原因和解决方案**

- 版本修改和字段长度修改

**影响范围**

**验证方案**

**涉及分支**

* master
* 600

**检查项**

- [ ] 需要更新依赖
- [ ] 日志打印不会造成磁盘占用空间过多
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 打包成功(如果有修改Makefile)

/assign @zhuofeng

+++ stream/jsonify/sql.go
@@ -36,7 +36,7 @@ func SQLLoggerColumns() []*ckdb.Column {
 		ckdb.NewColumn("exception_desc", ckdb.String).SetComment("异常描述"),
 
 		ckdb.NewColumn("duration", ckdb.UInt64).SetComment("响应时延, 请求报文与第一个响应报文的时长(us)"), // us
-		ckdb.NewColumn("affected_rows", ckdb.UInt8).SetComment("影响行数").SetIndex(ckdb.IndexNone),
+		ckdb.NewColumn("affected_rows", ckdb.UInt64).SetComment("影响行数").SetIndex(ckdb.IndexNone),
 	)
 	return columns
 }
@@ -94,7 +94,7 @@ func (s *SQLLogger) WriteBlock(block *ckdb.Block) error {
 			return err
 		}
 
-		if err := block.WriteUInt8(s.mysql.AffectedRows); err != nil {
+		if err := block.WriteUInt64(s.mysql.AffectedRows); err != nil {
 			return err
 		}
 	}
